% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BehrensRJags.init}
\alias{BehrensRJags.init}
\title{BehrensRJags.Init}
\usage{
BehrensRJags.init()
}
\description{
Retrieves a dependency package from Github that is not available on the CRAN repositories. This only needs to be run when BehrensRJags is first installed.
}
\details{
run.model(...) will not work without calling this function at least once to install the jagstools package. The model will build and update, but the output will break.
}
\examples{
BehrensRJags.init()
}

