% Generated by roxygen2 (4.0.1): do not edit by hand
\name{run.model}
\alias{run.model}
\title{run.model}
\usage{
run.model(volatility.fixed = FALSE, model.file, data.file, n.chains,
  n.samples, n.burnin)
}
\arguments{
\item{volatility.fixed}{Determines whether the volatility is fixed and stationary in time if true, or changes throughout time if false. Defaults to false}

\item{model.file}{The path to the text file containing the model}

\item{data.file}{The path to the csv file containing the data}

\item{n.chains}{The amount of Markov chains to run}

\item{n.samples}{The amount of samples to take. More samples = more convergence}

\item{n.burnin}{The amount of burnin samples to discard}
}
\description{
Runs Behren's model with R2jags and plots means of expecation and volatility through time.
}
\details{
Output is in CSV format. File names are as follows: \cr\cr <data.file.name>_ts_*.csv - The "_ts_" (time-series) indicates that these are the results for the samples of the time series variables \cr\cr <data.file.name>_s_*.csv - The "_s_" (stationary) indicates that these are the results for the samples of the time stationary variables \cr\cr <data.file.name>_*_v.csv - The "_v" (volatility time series) indicates that these are the results for the samples of the variables with v being a time series variable \cr\cr <data.file.name>_*_vf.csv - The "_vf" (volatility fixed) indicates that these are the results for the samples of the variables with v being a time stationary variable \cr\cr Columns are as follows: \cr\cr *_ts_v.csv - expectation mean, expectation standard deviation, volatility mean, volatility standard deviation \cr\cr *_s_v.csv - k mean, k standard deviation, expectation beta parameter mean, expectation beta parameter standard deviation, action beta parameter mean, action beta parameter standard deviation \cr\cr *_ts_vf.csv - expectation mean, expectation standard deviation \cr\cr *_s_vf.csv - volatility mean, volatility standard deviation, expectation beta parameter mean, expectation beta parameter standard deviation, action beta parameter mean, action beta parameter standard deviation
}
\examples{
run.model(model.file=".../model.txt", data.file=".../data/subj_005_social.csv", n.chains=1, n.samples=15000, n.burnin=1000)
run.model(volatility.fixed=TRUE, model.file=".../model-fixed-volatility.txt", data.file=".../data/subj_005_social.csv", n.chains=1, n.samples=15000, n.burnin=1000)
}
\keyword{behren}
\keyword{model}

