% Generated by roxygen2 (4.0.1): do not edit by hand
\name{runModel}
\alias{runModel}
\title{runModel}
\usage{
runModel(modelFile, dataFile, n.chains, n.samples, n.burnin)
}
\arguments{
\item{modelFile}{The path to the text file containing the model}

\item{dataFile}{The path to the csv file containing the data}

\item{n.chains}{The amount of Markov chains to run}

\item{n.samples}{The amount of samples to take. More samples = more convergence}

\item{n.burnin}{The amount of burnin samples to discard}
}
\description{
Runs Behren's model with R2jags and plots means of expecation and volatility through time.
}
\examples{
runModel(modelFile=".../model.txt", dataFile=".../data/subj_005_social.csv", n.chains=1, n.samples=15000, n.burnin=1000)
}
\keyword{behren}
\keyword{model}

