\name{BehrensRJags-package}
\alias{BehrensRJags-package}
\alias{BehrensRJags}
\docType{package}
\title{
Behren's Behavior Model
}
\description{
Provides tools to analyze Behren's behavior model with R and JAGS
}
\details{
\tabular{ll}{
Package: \tab BehrensRJags\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-14\cr
License: \tab UALR\cr
}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
Jeff Long, Keith Bush

Maintainer: Jeff Long <jmlong@ualr.edu>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
