\name{BehrensRJags-package}
\alias{BehrensRJags-package}
\alias{BehrensRJags}
\docType{package}
\title{
BehrensRJags
}
\description{
Provides wrapper tools to analyze Timothy Behren's behavior model with R and JAGS.
}
\details{
\tabular{ll}{
Package: \tab BehrensRJags\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-14\cr
License: \tab UALR\cr
}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
Jeff Long, Keith Bush

Maintainer: Jeff Long <jmlong@ualr.edu>
}
\references{
 - Behrens, Timothy E.J., Nat. Neurosci. Vol 10, 1214-1221, September 2007
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
